requirements.txt
# Vesper AI Project - Python Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Stripe
stripe==10.5.0

# Email
email-validator==2.1.0

# Database (SQLite is built-in, but using sqlite3)
# No additional packages needed for SQLite

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production Server
gunicorn==21.2.0
